segmentation/pipelines/requirements.txt
addict==2.4.0
aiobotocore==2.4.2
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
backcall==0.2.0
billiard==3.6.4.0
boto3==1.24.59
botocore==1.27.59
brotlipy==0.7.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.4
dulwich==0.21.3
dvc==2.45.1
dvc-data==0.40.3
dvc-http==2.30.2
dvc-objects==0.19.3
dvc-render==0.2.0
dvc-s3==2.21.0
dvc-studio-client==0.5.0
dvc-task==0.1.11
dvclive==2.1.0
executing==1.2.0
filelock==3.9.0
flatten-dict==0.4.2
flit_core==3.6.0
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
funcy==1.18
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
huggingface-hub==0.12.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==6.0.0
ipython==8.10.0
iterative-telemetry==0.0.7
jedi==0.18.2
jmespath==1.0.1
kiwisolver==1.4.4
kombu==5.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
#mkl-fft==1.3.1
#mkl-random==1.2.2
#mkl-service==2.4.0
mmcls==0.25.0
#mmcv==1.7.1
mmcv-full==1.7.1
mmsegmentation==0.30.0
model-index==0.1.11
multidict==6.0.4
nanotime==0.5.2
networkx==3.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python==4.7.0.68
openmim==0.3.6
ordered-set==4.1.0
orjson==3.8.6
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==21.3.1
prettytable==3.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
rich==13.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2023.1.0
s3transfer==0.6.0
scmrepo==0.1.11
setuptools==60.5.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sqltrie==0.0.28
stack-data==0.6.2
tabulate==0.9.0
timm==0.6.12
tomlkit==0.11.6
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
wheel==0.37.1
wrapt==1.14.1
yapf==0.32.0
yarl==1.8.2
zc.lockfile==2.0
zipp==3.12.1




